+++ code/react/src/events/events.ts
@@ -2,34 +2,13 @@
  * Global events subscribed to only once and consumed as [Observable] event producers.
  */
 
-/**
- * Based on [React.IMouseEvent]
- * Note:
- *    Repeated here as exporting this event via the Observables
- *    is causing Typescript type-build errors.
- */
-export type IMouseEvent = {
-  altKey: boolean;
-  button: number;
-  buttons: number;
-  clientX: number;
-  clientY: number;
-  ctrlKey: boolean;
-  metaKey: boolean;
-  movementX: number;
-  movementY: number;
-  pageX: number;
-  pageY: number;
-  relatedTarget: EventTarget;
-  screenX: number;
-  screenY: number;
-  shiftKey: boolean;
-};
-
 import { animationFrameScheduler, fromEvent as rxFromEvent, Observable, Subject } from 'rxjs';
 import { FromEventTarget } from 'rxjs/internal/observable/fromEvent';
 import { map, merge, observeOn, share } from 'rxjs/operators';
 import { is } from '../common';
+import { IMouseEvent, KeypressEvent, KepressObservable } from './types';
+
+export * from './types';
 
 const fromEvent = <T>(source: FromEventTarget<any> | undefined, event: string): Observable<T> => {
   return source ? rxFromEvent(source, event).pipe(share()) : new Subject<any>().pipe(share()); // NB: Safe when server-rendered.
@@ -41,7 +20,9 @@ const fromDocumentEvent = <T>(event: string): Observable<T> =>
 const fromWindowEvent = <T>(event: string): Observable<T> =>
   fromEvent(is.browser ? window : undefined, event);
 
-//
+/**
+ * Mouse events.
+ */
 export const click$ = fromDocumentEvent<IMouseEvent>('click');
 export const mouseDown$ = fromDocumentEvent<IMouseEvent>('mousedown');
 export const mouseUp$ = fromDocumentEvent<IMouseEvent>('mouseup');
@@ -56,20 +37,6 @@ export const resize$ = fromWindowEvent<{}>('resize').pipe(
 /**
  * Keyboard events.
  */
-export type KeypressEvent = {
-  isPressed: boolean;
-  char: string | null;
-  code: string;
-  charCode: number;
-  key: string;
-  altKey: boolean;
-  ctrlKey: boolean;
-  shiftKey: boolean;
-  metaKey: boolean;
-  isModifier: boolean;
-  preventDefault: () => void;
-};
-export type KepressObservable = Observable<KeypressEvent>;
 
 const toKeypress = (e: KeyboardEvent, isPressed: boolean) => {
   const { key, code, charCode, altKey, ctrlKey, shiftKey, metaKey, char } = e;
@@ -86,6 +53,8 @@ const toKeypress = (e: KeyboardEvent, isPressed: boolean) => {
     metaKey,
     isModifier,
     preventDefault: () => e.preventDefault(),
+    stopPropagation: () => e.stopPropagation(),
+    stopImmediatePropagation: () => e.stopImmediatePropagation(),
   };
   return event;
 };
