+++ code/react/src/index.ts
@@ -3,3 +3,4 @@ export * from './events';
 export * from './measure';
 export * from './mouse';
 export * from './util';
+export * from './types';
